Add Terraform configuration for creating GCS bucket
This commit adds the necessary Terraform configuration files to create a GCS bucket in a GCP project. The configuration includes variable definitions for the project ID, bucket name, region, and storage class, as well as a resource block for the GCS bucket. The README file has also been updated with instructions on how to use the Terraform configuration.